refactor/#385: onSubmitReview 로직 훅으로 분리
